Display live PASS events for packages that finished testing
This patch starts a goroutine that takes care of rendering events of particular
interest, like test stdio when -follow is specified and PASS lines by default.

This adds a form of interactivity for large/slow test suites, without having to
enable -follow, which could be extremely verbose for some code bases.